# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/discounts": {
            "get": {
                "description": "Retrieves a list of all discounts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Get all discounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/API.DiscountResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new discount by providing discount details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Create a new discount",
                "parameters": [
                    {
                        "description": "Discount details",
                        "name": "discount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/API.CreateDiscountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/discounts/paginated": {
            "get": {
                "description": "Retrieves a list of discounts with pagination and filtering options.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Get paginated list of discounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search value",
                        "name": "searchValue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page index",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Discount status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/API.DiscountResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/discounts/{id}": {
            "get": {
                "description": "Retrieves a discount by providing the discount ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Get discount by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.DiscountResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a discount by providing the discount ID and new discount details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Update discount by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated discount details",
                        "name": "discount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/API.UpdateDiscountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a discount by providing the discount ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discounts"
                ],
                "summary": "Delete discount by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/freightRates": {
            "get": {
                "description": "Retrieves a list of freight rates for couriers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Freight"
                ],
                "summary": "Get all freight rates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/API.FreightRate"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/freightRates/{id}": {
            "get": {
                "description": "Retrieves a specific freight rate by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Freight"
                ],
                "summary": "Get freight rate by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Freight Rate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.FreightRate"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "API.CreateDiscountRequest": {
            "type": "object",
            "required": [
                "discount_type",
                "discount_value",
                "end_date",
                "start_date"
            ],
            "properties": {
                "discount_type": {
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "API.DiscountResponse": {
            "type": "object",
            "properties": {
                "discount_type": {
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "API.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "API.FreightRate": {
            "type": "object",
            "properties": {
                "cost_per_km": {
                    "type": "number"
                },
                "courier_id": {
                    "type": "string"
                },
                "distance_max_km": {
                    "type": "integer"
                },
                "distance_min_km": {
                    "type": "integer"
                },
                "rate_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "API.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "API.UpdateDiscountRequest": {
            "type": "object",
            "required": [
                "discount_type",
                "discount_value",
                "end_date",
                "start_date"
            ],
            "properties": {
                "discount_type": {
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
    }
}